Store discriminator column in every foreign entity
This prevents the complexity and performance penalty of eager-loading the target entity for the sole purpose of knowing the target class name, at the expense of storing the discriminator value in every entity linking to an entity in an inheritance hierarchy.

The discriminator value can easily fit in a single byte, though, so the benefits should outweigh the costs.